헤더 푸터 컴포넌트 나눔
