+++ front-practice1/package.json
@@ -18,6 +18,7 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-modal": "^3.16.1",
+    "react-router-dom": "^6.24.0",
     "react-scripts": "5.0.1",
     "sass-loader": "^14.2.1",
     "style-loader": "^4.0.0",
