+++ front-practice1/src/Components/Container/Calendar.jsx
@@ -3,7 +3,7 @@ import FullCalendar from "@fullcalendar/react";
 import dayGridPlugin from "@fullcalendar/daygrid";
 import timeGridPlugin from "@fullcalendar/timegrid";
 import interactionPlugin from "@fullcalendar/interaction";
-import '../../Css/Calendar.css'
+import "../../Css/Calendar.css";
 
 function Calendar() {
     return(
@@ -12,8 +12,14 @@ function Calendar() {
                 defaultView="dayGridMonth" 
                 plugins={[ dayGridPlugin, timeGridPlugin, interactionPlugin ]}
                 events={[
-                    {title : '다이어리 1', date: '2024-06-25'},
-                    {title : '가계부 1', date: '2024-06-25'}
+                    {
+                        title : '다이어리',
+                        date: '2024-06-25'
+                    },
+                    {
+                        title : '가계부',
+                        date: '2024-06-25'
+                    }
                 ]}
                 height={"85vh"}
             />
