+++ front-practice1/src/Components/Footer/Footer.jsx
@@ -1,0 +1,11 @@
+
+
+export default function Footer() {
+    return (
+        <>
+            <footer>
+                <p>footer</p>
+            </footer>
+        </>
+    )
+}
